flask_app/requirement.txt
blinker==1.6.2
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
Flask==2.3.2
fonttools==4.39.4
freetype-py==2.3.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
multiprocessing-logging==0.3.4
ntplib==0.4.0
numpy==1.24.3
packaging==23.1
pandas==2.0.1
patsy==0.5.3
Pillow==9.5.0
psycopg2-binary==2.9.6
pycairo==1.23.0
pydicom==2.3.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
reportlab==4.0.0
rlPyCairo==0.2.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
scikit-learn
statsmodels==0.14.0
threadpoolctl==3.1.0
tzdata==2023.3
Werkzeug==2.3.4
zipp==3.15.0
openpyxl